Added move_utils.c
