+++ srcs/hook_utils.c
@@ -8,47 +8,14 @@ int	ft_close_window(t_cub *cub)
 
 void	ft_rotate(t_cub *cub, double angle)
 {
-	double	old_dir_x;
-	double	old_plane_x;
-
-	old_dir_x = cub->raycast->player_dir.x;
-	cub->raycast->player_dir.x = cub->raycast->player_dir.x * cos(angle) - cub->raycast->player_dir.y * sin(angle);
-	cub->raycast->player_dir.y = old_dir_x * sin(angle) + cub->raycast->player_dir.y * cos(angle);
-	old_plane_x = cub->raycast->camera_plane.x; 
-	cub->raycast->camera_plane.x = cub->raycast->camera_plane.x * cos(angle) - cub->raycast->camera_plane.y * sin(angle);
-	cub->raycast->camera_plane.y = old_plane_x * sin(angle) + cub->raycast->camera_plane.y * cos(angle);
-}
-
-void	ft_move_left(t_cub *cub, double *tmp_x, double *tmp_y)
-{
-	*tmp_x -= cub->raycast->camera_plane.x * cub->raycast->move_speed;
-	*tmp_y -= cub->raycast->camera_plane.y * cub->raycast->move_speed;
-}
-
-void	ft_move_right(t_cub *cub, double *tmp_x, double *tmp_y)
-{
-	*tmp_x += cub->raycast->camera_plane.x * cub->raycast->move_speed;
-	*tmp_y += cub->raycast->camera_plane.y * cub->raycast->move_speed;
-}
-
-void	ft_move_up(t_cub *cub, double *tmp_x, double *tmp_y)
-{
-	*tmp_x += cub->raycast->player_dir.x * cub->raycast->move_speed;
-	*tmp_y += cub->raycast->player_dir.y * cub->raycast->move_speed;
-}
-
-void	ft_move_down(t_cub *cub, double *tmp_x, double *tmp_y)
-{
-	*tmp_x -= cub->raycast->player_dir.x * cub->raycast->move_speed;
-	*tmp_y -= cub->raycast->player_dir.y * cub->raycast->move_speed;
-}
-
-void	ft_update_position(t_cub *cub, double tmp_x, double tmp_y)
-{
-	cub->raycast->player_pos.x = tmp_x;
-	cub->raycast->player_pos.y = tmp_y;
-	cub->raycast->player_squ.x = (int)tmp_x;
-	cub->raycast->player_squ.y = (int)tmp_y;
+	cub->raycast->player_dir.y = cub->raycast->player_dir.x * sin(angle)
+		+ cub->raycast->player_dir.y * cos(angle);
+	cub->raycast->player_dir.x = cub->raycast->player_dir.x * cos(angle)
+		- cub->raycast->player_dir.y * sin(angle);
+	cub->raycast->camera_plane.y = cub->raycast->camera_plane.x * sin(angle)
+		+ cub->raycast->camera_plane.y * cos(angle);
+	cub->raycast->camera_plane.x = cub->raycast->camera_plane.x * cos(angle)
+		- cub->raycast->camera_plane.y * sin(angle);
 }
 
 int	ft_key_input(int keysym, t_cub *cub)
